pyproject.toml
[tool.poetry]
name = "script-args-parser"
version = "0.4.0"
description = "Script arguments parsing library."
authors = ["KRunchPL <krunchfrompoland@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/KRunchPL/script-args-parser"
repository = "https://github.com/KRunchPL/script-args-parser"
documentation = "https://github.com/KRunchPL/script-args-parser"
license = "Apache-2.0"
packages = [
    { include = "script_args_parser" },
]

[tool.poetry.dependencies]
python = "^3.9"
toml = "^0.10.2"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
darglint = "^1.8.1"
flake8 = "^3.8.4"
flake8-html = "^0.4.1"
flake8-isort = "^4.1.1"
flake8-pydocstyle = "^0.2.0"
flake8-todo = "^0.7"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
mypy = "^0.910"
types-toml = "^0.10.1"
types-PyYAML = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
addopts = "--cov script_args_parser --cov-report html:coverage-report"

[tool.isort]
virtual_env = ".venv"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
lines_between_types = 0
lines_between_sections = 1
lines_after_imports = 2
lines_before_imports = 0
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3
line_length = 110
force_grid_wrap = 0
combine_as_imports = true
ensure_newline_before_comments = true

[tool.pydocstyle]
### Ignoring explanation
# D104 - Missing docstring in public package
# D107 - Missing docstring in __init__
# D200 - One-line docstring should fit on one line with quotes
# D203 - 1 blank line required before class docstring
# D212 - Multi-line docstring summary should start at the first line
ignore = """
D104,D107,
D200,D203,D212
"""

[tool.mypy]
files = "."
exclude = "(^|/)tests/"
# Strict options
strict = true
warn_unused_configs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true
# Additional options
disallow_untyped_globals = true
warn_no_return = true
warn_unreachable = true
